+++ projects/ngx-opendesign/src/public-api.ts
@@ -40,6 +40,7 @@ export * from './lib/layout/container';
 export * from './lib/layout/stack';
 export * from './lib/layout/grid';
 export * from './lib/layout/layout';
+export * from './lib/splitter/splitter';
 export * from './lib/table/table';
 export * from './lib/table/table-container';
 export * from './lib/table/sort-header';